whisper-tts/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gtts==2.4.0
requests==2.31.0
werkzeug==3.0.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
